Conversion to TypeScript compiling
